+++ src/main.rs
@@ -1,14 +1,15 @@
 #![windows_subsystem = "windows"]
-use gtk::prelude::*;
-use gtk::{gio, glib};
 mod config;
 mod ui;
+
+use config::load_uuid;
+use gtk::prelude::*;
+use gtk::{gio, glib};
 
 const APP_ID: &str = "moe.tsuna.tsukimi";
 
 fn main() -> glib::ExitCode {
-    // Load the configuration file
-    config::load_cfg();
+    load_uuid();
     // Register and include resources
     gio::resources_register_include!("tsukimi.gresource").expect("Failed to register resources.");