+++ src/ui/moviedrop.rs
@@ -1,9 +1,11 @@
+use gtk::prelude::*;
+
+use super::models::SETTINGS;
 use super::network;
 use super::network::Back;
 use super::network::SearchResult;
 use super::new_dropsel::play_event;
 use super::provider::dropdown_factory::factory;
-use gtk::prelude::*;
 
 pub fn newmediadropsel(
     playbackinfo: network::Media,
@@ -61,8 +63,7 @@ pub fn newmediadropsel(
     });
 
     let info = info.clone();
-    let settings = gtk::gio::Settings::new(crate::APP_ID);
-    if settings.boolean("is-resume") {
+    if SETTINGS.resume() {
         if let Some(userdata) = &info.user_data {
             if let Some(percentage) = userdata.played_percentage {
                 if percentage > 0. {