+++ src/ui/widgets/actor.rs
@@ -1,9 +1,10 @@
-use crate::ui::image::setimage;
 use adw::prelude::NavigationPageExt;
 use glib::Object;
 use gtk::prelude::*;
 use gtk::subclass::prelude::*;
 use gtk::{gio, glib};
+
+use crate::ui::image::setimage;
 
 use super::fix::fix;
 use super::item::ItemPage;
@@ -82,9 +83,25 @@ mod imp {
         fn constructed(&self) {
             self.parent_constructed();
             let obj = self.obj();
-            obj.setup_pic();
-            obj.get_item();
-            obj.set_lists();
+            let (sender, receiver) = async_channel::bounded::<bool>(1);
+            gtk::gio::spawn_blocking(move || {
+                sender
+                    .send_blocking(false)
+                    .expect("The channel needs to be open.");
+                std::thread::sleep(std::time::Duration::from_millis(400));
+                sender
+                    .send_blocking(true)
+                    .expect("The channel needs to be open.");
+            });
+            glib::spawn_future_local(glib::clone!(@weak obj =>async move {
+                while let Ok(bool) = receiver.recv().await {
+                    if bool {
+                        obj.setup_pic();
+                        obj.get_item();
+                        obj.set_lists();
+                    }
+                }
+            }));
         }
     }