+++ src/ui/widgets/home.rs
@@ -1,10 +1,9 @@
-use std::env;
-
 use adw::prelude::NavigationPageExt;
 use glib::Object;
 use gtk::prelude::*;
 use gtk::subclass::prelude::*;
 use gtk::{gio, glib};
+use std::env;
 
 use crate::ui::network::Latest;
 
@@ -60,8 +59,24 @@ mod imp {
         fn constructed(&self) {
             self.parent_constructed();
             let obj = self.obj();
-            // request library
-            obj.set_library();
+            let (sender, receiver) = async_channel::bounded::<bool>(1);
+            gtk::gio::spawn_blocking(move || {
+                sender
+                    .send_blocking(false)
+                    .expect("The channel needs to be open.");
+                std::thread::sleep(std::time::Duration::from_millis(400));
+                sender
+                    .send_blocking(true)
+                    .expect("The channel needs to be open.");
+            });
+            glib::spawn_future_local(glib::clone!(@weak obj =>async move {
+                while let Ok(bool) = receiver.recv().await {
+                    if bool {
+                        // request library
+                        obj.set_library();
+                    }
+                }
+            }));
         }
     }