+++ src/ui/widgets/list.rs
@@ -1,20 +1,20 @@
-use super::item::ItemPage;
-use super::movie::MoviePage;
-use super::window::Window;
 use adw::prelude::NavigationPageExt;
 use glib::Object;
 use gtk::prelude::*;
 use gtk::subclass::prelude::*;
 use gtk::{gio, glib};
 
-mod imp {
+use super::item::ItemPage;
+use super::movie::MoviePage;
+use super::window::Window;
 
-    use std::cell::OnceCell;
+mod imp {
 
     use glib::subclass::InitializingObject;
     use gtk::prelude::*;
     use gtk::subclass::prelude::*;
     use gtk::{glib, CompositeTemplate};
+    use std::cell::OnceCell;
 
     // Object holding the state
     #[derive(CompositeTemplate, Default, glib::Properties)]
@@ -59,7 +59,23 @@ mod imp {
         fn constructed(&self) {
             self.parent_constructed();
             let obj = self.obj();
-            obj.set_factory();
+            let (sender, receiver) = async_channel::bounded::<bool>(1);
+            gtk::gio::spawn_blocking(move || {
+                sender
+                    .send_blocking(false)
+                    .expect("The channel needs to be open.");
+                std::thread::sleep(std::time::Duration::from_millis(500));
+                sender
+                    .send_blocking(true)
+                    .expect("The channel needs to be open.");
+            });
+            glib::spawn_future_local(glib::clone!(@weak obj =>async move {
+                while let Ok(bool) = receiver.recv().await {
+                    if bool {
+                        obj.set_factory();
+                    }
+                }
+            }));
         }
     }